Tests/LibWeb/Text/expected/abspos-relayout-rtl-static-position.txt

live=180,10,120,40
reference=180,10,120,40
partial-layouts=0 full-layouts=1
FAIL
